bot/requirements.txt
python-telegram-bot[job-queue]==21.0
psycopg2-binary==2.9.9
httpx>=0.24.0
